server: rewrite message handling
Currently we don't do any synchronization when we send out
network messages, this can cause a variety of issues. Instead
of adding serialization for transmits, let the messages queue
up in the network backend, and let the backend send them out.

Work in progress, not everything has been tested with this.

Signed-off-by: Jens Axboe <axboe@fb.com>