backend: continue on failed wait_for() check, not break
If we have this job file:

[a]

[b]
wait_for=a

[c]
wait_for=a

[d]

then we expect a/d to run together, then b/c after a is done. However,
right now we break out of the job loop, so it acts like a stonewall.
That means that a will run first, then b/c/d. That isn't correct.

Signed-off-by: Jens Axboe <axboe@fb.com>