Fix issue with termination before io_size has been reached
If the size of the device isn't a multiple of the block size,
then we can exit prematurely for cases where an io_limit has
been set higher than the device size. This is because the file(s)
get marked completely done if bs is a multiple of the device size,
but doesn't for the opposite case.

We should not factor this in for deciding to keep running or not.

Signed-off-by: Jens Axboe <axboe@fb.com>