options: full control of fadvise hinting
Previously fio supported only 0/1 for this hint. 0 would not do any
hints, but 1 (the default) would hint based on what the fio workload
was - fio would tell the kernel it was a random workload, if it was
random, or a sequential for a sequential workload.

Now fio also supports setting what the hint type should be explicitly,
by allowing a setting of 'random' or 'sequential' for full control
of the hint.

Signed-off-by: Jens Axboe <axboe@fb.com>