fio: warn if user asks for iodepth that can't be reached
Most synchronous I/O engines can only submit one I/O at a time and have
to wait for that I/O to complete before submitting more I/O. However
users often don't realise this and ask for an iodepth that can never be
achieved.

Introduce a new engine flag that states whether an I/O engine can submit
multiple I/Os at once and automatically set it if an I/O engine is
asynchronous. Add a warning to check if the user asked for an iodepth
which we know can't be reached but don't warn if:
- io_submit_mode is offload (synchronous I/O will be submitted from
  iodepth threads/processes)
- verify_async is set (additional synchronous I/O will be submitted from
  verify_async threads/processes)

Signed-off-by: Sitsofe Wheeler <sitsofe@yahoo.com>