stat: rename JSON latency_ns/us/ms objects to clatency_ns/us/ms
The JSON output has latency_ns/us/ms objects that describe the *completion*
latency distribution for all of the IOs in a job. Since the normal output
uses slat for submission, clat for completion, and lat for total latencies,
it seems appropriate to rename these to clatency_ns/us/ms. This will of course
break automated parsing of these JSON objects, but it more accurately describes
what these values mean.