rate: fix bad math
To figure out how much data we missed when doing a thinktime sleep,
we're currently dividing by the time slept. This is wrong, it should
be multiplied by the time slept and divided by 1000000 to go from
usec to a second base.

Additionally, don't ever subtract more than a block of data, and
adjust down depending on sleep.

Fixes: 1aa39b0c ("rate: ensure IO issue restarts right after sleep")
Signed-off-by: Jens Axboe <axboe@kernel.dk>